+++ application/languages/en/non_spdx_licenses.php
@@ -1,0 +1,7 @@
+<?php
+return array(
+        "NOASSERTION-HAL-Authorisation-1.0" => "Hal authorisation v1",
+        "NOASSERTION-HAL-Copyright" => "Hal Copyright",
+        "NOASSERTION-ARXIV-1.0-nonexclusive-distrib" => "arXiv.org - Non-exclusive license to distribute",
+        "NOASSERTION-ARXIV-assumed-1991-2003" => "arXiv.org - Assumed non-exclusive license to distribute"
+);