+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 ## Unreleased
 ### Added
-### Added
+- [#893](https://github.com/CCSDForge/episciences/issues/893) Allow to select the license of the document use spdx.
 - [#630](https://github.com/CCSDForge/episciences/issues/630) COI (Conflict of Interest) notifications:
     - Email notification to the editor-in-chief when an editor declares a positive COI (answered "yes")
     - Email notification to other assigned editors (if any) when an editor declares a COI